# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperror.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
    type: object
  handler.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  handler.ListUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.UserResponse'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Password1!
        type: string
    type: object
  handler.ReadyzResponse:
    properties:
      db:
        example: true
        type: boolean
      redis:
        example: true
        type: boolean
      status:
        example: ready
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      email:
        example: newuser@example.com
        type: string
      password:
        example: Password1!
        type: string
    type: object
  handler.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
  handler.TokenResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
    type: object
  handler.healthResponse:
    properties:
      status:
        type: string
    type: object
  model.UpdateUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      role:
        type: string
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: api@example.com
    name: API Support
    url: https://github.com/nowi5/kleido
  description: Production-grade Go REST API. JWT RS256 authentication required on
    all /api/v1/ endpoints except /auth/register and /auth/login.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kleido API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: |-
        Sends a password reset email. Always returns 200 regardless of whether
        the email is registered, to prevent account enumeration.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: if your email is registered, a reset link has been
            sent'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns an access token. The refresh
        token is set as an httpOnly cookie.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "400":
          description: Missing or malformed request body
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      summary: Authenticate a user
      tags:
      - auth
  /auth/logout:
    post:
      description: Blocklists the access token's JTI and revokes the refresh token.
        Clears the refresh_token cookie.
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke the current session
      tags:
      - auth
  /auth/refresh:
    post:
      description: Issues a new access token using the refresh_token cookie. The old
        refresh token is revoked and replaced with a new one (rotation).
      produces:
      - application/json
      responses:
        "200":
          description: New access token; new refresh_token cookie set
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "401":
          description: Missing or invalid refresh token cookie
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the "user" role.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: |-
        Sets a new password using a valid reset token from the email link.
        The token is single-use and expires after 1 hour.
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: password updated'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid/expired token or weak password
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      summary: Reset password using token
      tags:
      - auth
  /healthz:
    get:
      description: Returns 200 if the process is running. No downstream checks.
      produces:
      - application/json
      responses:
        "200":
          description: 'status: ok'
          schema:
            $ref: '#/definitions/handler.healthResponse'
      summary: Liveness probe
      tags:
      - system
  /readyz:
    get:
      description: Returns 200 when PostgreSQL and Redis are reachable. Returns 503
        if either is down.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ReadyzResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handler.ReadyzResponse'
      summary: Readiness probe
      tags:
      - system
  /users:
    get:
      description: Returns a paginated list of all users. Admin only.
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft-deletes a user (sets is_active = false). Admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Returns a single active user by UUID. Admins can see all users;
        regular users can only fetch their own profile.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Non-admin fetching another user
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Partially updates a user's email, role, or active status. Role
        changes require admin privileges; non-admins may only update their own profile.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (all optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Non-admin updating another user
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "409":
          description: Email already taken
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/me:
    get:
      description: Returns the profile of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the authenticated user's profile
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT access token. Example:
      "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"